+++ sentry-android-core/src/main/java/io/sentry/android/core/AndroidTransactionProfiler.java
@@ -7,78 +7,39 @@ import android.annotation.SuppressLint;
 import android.app.ActivityManager;
 import android.content.Context;
 import android.os.Build;
-import android.os.Debug;
 import android.os.Process;
 import android.os.SystemClock;
-import io.sentry.CpuCollectionData;
 import io.sentry.HubAdapter;
 import io.sentry.IHub;
 import io.sentry.ITransaction;
 import io.sentry.ITransactionProfiler;
-import io.sentry.MemoryCollectionData;
 import io.sentry.PerformanceCollectionData;
 import io.sentry.ProfilingTraceData;
 import io.sentry.ProfilingTransactionData;
 import io.sentry.SentryLevel;
 import io.sentry.android.core.internal.util.CpuInfoUtils;
 import io.sentry.android.core.internal.util.SentryFrameMetricsCollector;
-import io.sentry.profilemeasurements.ProfileMeasurement;
-import io.sentry.profilemeasurements.ProfileMeasurementValue;
 import io.sentry.util.Objects;
-import java.io.File;
-import java.util.ArrayDeque;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.UUID;
-import java.util.concurrent.Future;
-import java.util.concurrent.RejectedExecutionException;
-import java.util.concurrent.TimeUnit;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.TestOnly;
 
 final class AndroidTransactionProfiler implements ITransactionProfiler {
-
-  /**
-   * This appears to correspond to the buffer size of the data part of the file, excluding the key
-   * part. Once the buffer is full, new records are ignored, but the resulting trace file will be
-   * valid.
-   *
-   * <p>30 second traces can require a buffer of a few MB. 8MB is the default buffer size for
-   * [Debug.startMethodTracingSampling], but 3 should be enough for most cases. We can adjust this
-   * in the future if we notice that traces are being truncated in some applications.
-   */
-  private static final int BUFFER_SIZE_BYTES = 3_000_000;
-
-  private static final int PROFILING_TIMEOUT_MILLIS = 30_000;
-
-  private int intervalUs;
-  private @Nullable File traceFile = null;
-  private @Nullable File traceFilesDir = null;
-  private @Nullable Future<?> scheduledFinish = null;
-  private volatile @Nullable ProfilingTraceData timedOutProfilingData = null;
   private final @NotNull Context context;
   private final @NotNull SentryAndroidOptions options;
   private final @NotNull IHub hub;
   private final @NotNull BuildInfoProvider buildInfoProvider;
-  private long transactionStartNanos = 0;
-  private long profileStartCpuMillis = 0;
   private boolean isInitialized = false;
   private int transactionsCounter = 0;
-  private @Nullable String frameMetricsCollectorId;
   private final @NotNull SentryFrameMetricsCollector frameMetricsCollector;
   private @Nullable ProfilingTransactionData currentProfilingTransactionData;
-  private final @NotNull ArrayDeque<ProfileMeasurementValue> screenFrameRateMeasurements =
-      new ArrayDeque<>();
-  private final @NotNull ArrayDeque<ProfileMeasurementValue> slowFrameRenderMeasurements =
-      new ArrayDeque<>();
-  private final @NotNull ArrayDeque<ProfileMeasurementValue> frozenFrameRenderMeasurements =
-      new ArrayDeque<>();
-  private final @NotNull Map<String, ProfileMeasurement> measurementsMap = new HashMap<>();
 
   private @Nullable ITransaction currentTransaction = null;
+  private @Nullable AndroidProfiler profiler = null;
+  private long transactionStartNanos;
+  private long profileStartCpuMillis;
 
   public AndroidTransactionProfiler(
       final @NotNull Context context,
@@ -137,8 +98,14 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
               intervalHz);
       return;
     }
-    intervalUs = (int) SECONDS.toMicros(1) / intervalHz;
-    traceFilesDir = new File(tracesFilesDirPath);
+
+    profiler =
+        new AndroidProfiler(
+            tracesFilesDirPath,
+            (int) SECONDS.toMicros(1) / intervalHz,
+            frameMetricsCollector,
+            options,
+            buildInfoProvider);
   }
 
   @Override
@@ -148,12 +115,6 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
 
     // Let's initialize trace folder and profiling interval
     init();
-
-    // traceFilesDir is null or intervalUs is 0 only if there was a problem in the init, but
-    // we already logged that
-    if (traceFilesDir == null || intervalUs == 0) {
-      return;
-    }
 
     transactionsCounter++;
     // When the first transaction is starting, we can start profiling
@@ -181,97 +142,24 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
 
   @SuppressLint("NewApi")
   private boolean onFirstTransactionStarted(final @NotNull ITransaction transaction) {
-    // We create a file with a uuid name, so no need to check if it already exists
-    traceFile = new File(traceFilesDir, UUID.randomUUID() + ".trace");
+    // init() didn't create profiler, should never happen
+    if (profiler == null) {
+      return false;
+    }
 
-    measurementsMap.clear();
-    screenFrameRateMeasurements.clear();
-    slowFrameRenderMeasurements.clear();
-    frozenFrameRenderMeasurements.clear();
-
-    frameMetricsCollectorId =
-        frameMetricsCollector.startCollection(
-            new SentryFrameMetricsCollector.FrameMetricsCollectorListener() {
-              final long nanosInSecond = TimeUnit.SECONDS.toNanos(1);
-              final long frozenFrameThresholdNanos = TimeUnit.MILLISECONDS.toNanos(700);
-              float lastRefreshRate = 0;
-
-              @Override
-              public void onFrameMetricCollected(
-                  final long frameEndNanos, final long durationNanos, float refreshRate) {
-                // transactionStartNanos is calculated through SystemClock.elapsedRealtimeNanos(),
-                // but frameEndNanos uses System.nanotime(), so we convert it to get the timestamp
-                // relative to transactionStartNanos
-                final long frameTimestampRelativeNanos =
-                    frameEndNanos
-                        - System.nanoTime()
-                        + SystemClock.elapsedRealtimeNanos()
-                        - transactionStartNanos;
-
-                // We don't allow negative relative timestamps.
-                // So we add a check, even if this should never happen.
-                if (frameTimestampRelativeNanos < 0) {
-                  return;
-                }
-                // Most frames take just a few nanoseconds longer than the optimal calculated
-                // duration.
-                // Therefore we subtract one, because otherwise almost all frames would be slow.
-                boolean isSlow = durationNanos > nanosInSecond / (refreshRate - 1);
-                float newRefreshRate = (int) (refreshRate * 100) / 100F;
-                if (durationNanos > frozenFrameThresholdNanos) {
-                  frozenFrameRenderMeasurements.addLast(
-                      new ProfileMeasurementValue(frameTimestampRelativeNanos, durationNanos));
-                } else if (isSlow) {
-                  slowFrameRenderMeasurements.addLast(
-                      new ProfileMeasurementValue(frameTimestampRelativeNanos, durationNanos));
-                }
-                if (newRefreshRate != lastRefreshRate) {
-                  lastRefreshRate = newRefreshRate;
-                  screenFrameRateMeasurements.addLast(
-                      new ProfileMeasurementValue(frameTimestampRelativeNanos, newRefreshRate));
-                }
-              }
-            });
+    final AndroidProfiler.ProfileStartData startData = profiler.start();
+    // check if profiling started
+    if (startData == null) {
+      return false;
+    }
+    transactionStartNanos = startData.startNanos;
+    profileStartCpuMillis = startData.startCpuMillis;
 
     currentTransaction = transaction;
 
-    // We stop profiling after a timeout to avoid huge profiles to be sent
-    try {
-      scheduledFinish =
-          options
-              .getExecutorService()
-              .schedule(
-                  () -> timedOutProfilingData = onTransactionFinish(transaction, true, null),
-                  PROFILING_TIMEOUT_MILLIS);
-    } catch (RejectedExecutionException e) {
-      options
-          .getLogger()
-          .log(
-              SentryLevel.ERROR,
-              "Failed to call the executor. Profiling will not be automatically finished. Did you call Sentry.close()?",
-              e);
-    }
-
-    transactionStartNanos = SystemClock.elapsedRealtimeNanos();
-    profileStartCpuMillis = Process.getElapsedCpuTime();
-
     currentProfilingTransactionData =
         new ProfilingTransactionData(transaction, transactionStartNanos, profileStartCpuMillis);
-
-    // We don't make any check on the file existence or writeable state, because we don't want to
-    // make file IO in the main thread.
-    // We cannot offload the work to the executorService, as if that's very busy, profiles could
-    // start/stop with a lot of delay and even cause ANRs.
-    try {
-      // If there is any problem with the file this method will throw (but it will not throw in
-      // tests)
-      Debug.startMethodTracingSampling(traceFile.getPath(), BUFFER_SIZE_BYTES, intervalUs);
-      return true;
-    } catch (Throwable e) {
-      onTransactionFinish(transaction, null);
-      options.getLogger().log(SentryLevel.ERROR, "Unable to start a profile: ", e);
-      return false;
-    }
+    return true;
   }
 
   @Override
@@ -287,34 +175,18 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
       final @NotNull ITransaction transaction,
       final boolean isTimeout,
       final @Nullable List<PerformanceCollectionData> performanceCollectionData) {
+    // check if profiler was created
+    if (profiler == null) {
+      return null;
+    }
 
     // onTransactionStart() is only available since Lollipop
     // and SystemClock.elapsedRealtimeNanos() since Jelly Bean
     if (buildInfoProvider.getSdkInfoVersion() < Build.VERSION_CODES.LOLLIPOP) return null;
 
-    final ProfilingTraceData profilingData = timedOutProfilingData;
-
     // Transaction finished, but it's not in the current profile
     if (currentProfilingTransactionData == null
         || !currentProfilingTransactionData.getId().equals(transaction.getEventId().toString())) {
-      // We check if we cached a profiling data due to a timeout with this profile in it
-      // If so, we return it back, otherwise we would simply lose it
-      if (profilingData != null) {
-        if (profilingData.getTransactionId().equals(transaction.getEventId().toString())) {
-          timedOutProfilingData = null;
-          return profilingData;
-        } else {
-          // Another transaction is finishing before the timed out one
-          options
-              .getLogger()
-              .log(
-                  SentryLevel.INFO,
-                  "A timed out profiling data exists, but the finishing transaction %s (%s) is not part of it",
-                  transaction.getName(),
-                  transaction.getSpanContext().getTraceId().toString());
-          return null;
-        }
-      }
       // A transaction is finishing, but it's not profiled. We can skip it
       options
           .getLogger()
@@ -338,7 +210,7 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
             transaction.getName(),
             transaction.getSpanContext().getTraceId().toString());
 
-    if (transactionsCounter != 0 && !isTimeout) {
+    if (transactionsCounter != 0) {
       // We notify the data referring to this transaction that it finished
       if (currentProfilingTransactionData != null) {
         currentProfilingTransactionData.notifyFinish(
@@ -350,19 +222,14 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
       return null;
     }
 
-    try {
-      // If there is any problem with the file this method could throw, but the start is also
-      // wrapped, so this should never happen (except for tests, where this is the only method that
-      // throws)
-      Debug.stopMethodTracing();
-    } catch (Throwable e) {
-      options.getLogger().log(SentryLevel.ERROR, "Error while stopping profiling: ", e);
+    final AndroidProfiler.ProfileEndData endData =
+        profiler.endAndCollect(false, performanceCollectionData);
+    // check if profiler end successfully
+    if (endData == null) {
+      return null;
     }
-    frameMetricsCollector.stopCollection(frameMetricsCollectorId);
 
-    long transactionEndNanos = SystemClock.elapsedRealtimeNanos();
-    long transactionEndCpuMillis = Process.getElapsedCpuTime();
-    long transactionDurationNanos = transactionEndNanos - transactionStartNanos;
+    long transactionDurationNanos = endData.endNanos - transactionStartNanos;
 
     List<ProfilingTransactionData> transactionList = new ArrayList<>(1);
     final ProfilingTransactionData txData = currentProfilingTransactionData;
@@ -373,16 +240,6 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
     // We clear the counter in case of a timeout
     transactionsCounter = 0;
     currentTransaction = null;
-
-    if (scheduledFinish != null) {
-      scheduledFinish.cancel(true);
-      scheduledFinish = null;
-    }
-
-    if (traceFile == null) {
-      options.getLogger().log(SentryLevel.ERROR, "Trace file does not exists");
-      return null;
-    }
 
     String totalMem = "0";
     ActivityManager.MemoryInfo memInfo = getMemInfo();
@@ -395,34 +252,13 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
     // Some may not have been really finished, in case of a timeout
     for (ProfilingTransactionData t : transactionList) {
       t.notifyFinish(
-          transactionEndNanos,
-          transactionStartNanos,
-          transactionEndCpuMillis,
-          profileStartCpuMillis);
+          endData.endNanos, transactionStartNanos, endData.endCpuMillis, profileStartCpuMillis);
     }
-
-    if (!slowFrameRenderMeasurements.isEmpty()) {
-      measurementsMap.put(
-          ProfileMeasurement.ID_SLOW_FRAME_RENDERS,
-          new ProfileMeasurement(ProfileMeasurement.UNIT_NANOSECONDS, slowFrameRenderMeasurements));
-    }
-    if (!frozenFrameRenderMeasurements.isEmpty()) {
-      measurementsMap.put(
-          ProfileMeasurement.ID_FROZEN_FRAME_RENDERS,
-          new ProfileMeasurement(
-              ProfileMeasurement.UNIT_NANOSECONDS, frozenFrameRenderMeasurements));
-    }
-    if (!screenFrameRateMeasurements.isEmpty()) {
-      measurementsMap.put(
-          ProfileMeasurement.ID_SCREEN_FRAME_RATES,
-          new ProfileMeasurement(ProfileMeasurement.UNIT_HZ, screenFrameRateMeasurements));
-    }
-    putPerformanceCollectionDataInMeasurements(performanceCollectionData);
 
     // cpu max frequencies are read with a lambda because reading files is involved, so it will be
     // done in the background when the trace file is read
     return new ProfilingTraceData(
-        traceFile,
+        endData.traceFile,
         transactionList,
         transaction,
         Long.toString(transactionDurationNanos),
@@ -437,86 +273,22 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
         options.getProguardUuid(),
         options.getRelease(),
         options.getEnvironment(),
-        isTimeout
+        (endData.didTimeout || isTimeout)
             ? ProfilingTraceData.TRUNCATION_REASON_TIMEOUT
             : ProfilingTraceData.TRUNCATION_REASON_NORMAL,
-        measurementsMap);
-  }
-
-  @SuppressLint("NewApi")
-  private void putPerformanceCollectionDataInMeasurements(
-      final @Nullable List<PerformanceCollectionData> performanceCollectionData) {
-
-    // onTransactionStart() is only available since Lollipop
-    // and SystemClock.elapsedRealtimeNanos() since Jelly Bean
-    if (buildInfoProvider.getSdkInfoVersion() < Build.VERSION_CODES.LOLLIPOP) {
-      return;
-    }
-
-    // This difference is required, since the PerformanceCollectionData timestamps are expressed in
-    // terms of System.currentTimeMillis() and measurements timestamps require the nanoseconds since
-    // the beginning, expressed with SystemClock.elapsedRealtimeNanos()
-    long timestampDiff =
-        SystemClock.elapsedRealtimeNanos()
-            - transactionStartNanos
-            - TimeUnit.MILLISECONDS.toNanos(System.currentTimeMillis());
-    if (performanceCollectionData != null) {
-      final @NotNull ArrayDeque<ProfileMeasurementValue> memoryUsageMeasurements =
-          new ArrayDeque<>(performanceCollectionData.size());
-      final @NotNull ArrayDeque<ProfileMeasurementValue> nativeMemoryUsageMeasurements =
-          new ArrayDeque<>(performanceCollectionData.size());
-      final @NotNull ArrayDeque<ProfileMeasurementValue> cpuUsageMeasurements =
-          new ArrayDeque<>(performanceCollectionData.size());
-      for (PerformanceCollectionData performanceData : performanceCollectionData) {
-        CpuCollectionData cpuData = performanceData.getCpuData();
-        MemoryCollectionData memoryData = performanceData.getMemoryData();
-        if (cpuData != null) {
-          cpuUsageMeasurements.add(
-              new ProfileMeasurementValue(
-                  TimeUnit.MILLISECONDS.toNanos(cpuData.getTimestampMillis()) + timestampDiff,
-                  cpuData.getCpuUsagePercentage()));
-        }
-        if (memoryData != null && memoryData.getUsedHeapMemory() > -1) {
-          memoryUsageMeasurements.add(
-              new ProfileMeasurementValue(
-                  TimeUnit.MILLISECONDS.toNanos(memoryData.getTimestampMillis()) + timestampDiff,
-                  memoryData.getUsedHeapMemory()));
-        }
-        if (memoryData != null && memoryData.getUsedNativeMemory() > -1) {
-          nativeMemoryUsageMeasurements.add(
-              new ProfileMeasurementValue(
-                  TimeUnit.MILLISECONDS.toNanos(memoryData.getTimestampMillis()) + timestampDiff,
-                  memoryData.getUsedNativeMemory()));
-        }
-      }
-      if (!cpuUsageMeasurements.isEmpty()) {
-        measurementsMap.put(
-            ProfileMeasurement.ID_CPU_USAGE,
-            new ProfileMeasurement(ProfileMeasurement.UNIT_PERCENT, cpuUsageMeasurements));
-      }
-      if (!memoryUsageMeasurements.isEmpty()) {
-        measurementsMap.put(
-            ProfileMeasurement.ID_MEMORY_FOOTPRINT,
-            new ProfileMeasurement(ProfileMeasurement.UNIT_BYTES, memoryUsageMeasurements));
-      }
-      if (!nativeMemoryUsageMeasurements.isEmpty()) {
-        measurementsMap.put(
-            ProfileMeasurement.ID_MEMORY_NATIVE_FOOTPRINT,
-            new ProfileMeasurement(ProfileMeasurement.UNIT_BYTES, nativeMemoryUsageMeasurements));
-      }
-    }
+        endData.measurementsMap);
   }
 
   @Override
   public void close() {
-    // we cancel any scheduled work
-    if (scheduledFinish != null) {
-      scheduledFinish.cancel(true);
-      scheduledFinish = null;
-    }
     // we stop profiling
     if (currentTransaction != null) {
       onTransactionFinish(currentTransaction, true, null);
+    }
+
+    // we have to first stop profiling otherwise we would lost the last profile
+    if (profiler != null) {
+      profiler.close();
     }
   }